requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
docstring-parser==0.15
Flask==3.0.0
google-api-core==2.14.0
google-auth==2.23.4
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.54.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonify==0.5
jsonref==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
networkx==3.2.1
numpy==1.26.2
pathlib==1.0.1
Pillow==10.1.0
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.9
Pygments==2.17.2
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
rsa==4.9
sievedata==0.1.0.0.post8
six==1.16.0
sseclient==0.0.27
tqdm==4.64.1
typeguard==4.1.5
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
youtube-dl==2021.12.17
